Basic code quality of Live View
