+++ dbms/src/Storages/StorageLiveView.cpp
@@ -9,6 +9,7 @@ distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License. */
+
 #include <Parsers/ASTSelectQuery.h>
 #include <Parsers/ASTCreateQuery.h>
 #include <Parsers/ASTWatchQuery.h>
@@ -17,10 +18,17 @@ limitations under the License. */
 #include <Parsers/ASTLiteral.h>
 #include <Interpreters/Context.h>
 #include <Interpreters/InterpreterDropQuery.h>
+#include <Interpreters/InterpreterSelectQuery.h>
 #include <DataStreams/NullBlockInputStream.h>
 #include <DataStreams/LiveViewBlockInputStream.h>
 #include <DataStreams/LiveViewEventsBlockInputStream.h>
 #include <DataStreams/MaterializingBlockInputStream.h>
+#include <DataStreams/IBlockOutputStream.h>
+#include <DataStreams/OneBlockInputStream.h>
+#include <DataStreams/BlocksBlockInputStream.h>
+#include <DataStreams/MaterializingBlockInputStream.h>
+#include <DataStreams/SquashingBlockInputStream.h>
+#include <DataStreams/copyData.h>
 #include <Common/typeid_cast.h>
 
 #include <Storages/StorageLiveView.h>
@@ -97,6 +105,174 @@ static void checkAllowedQueries(const ASTSelectQuery & query)
         checkAllowedQueries(inner_query->as<ASTSelectQuery &>());
     }
 }
+
+
+class LiveViewBlockOutputStream : public IBlockOutputStream
+{
+public:
+    explicit LiveViewBlockOutputStream(StorageLiveView & storage_) : storage(storage_) {}
+
+    void writePrefix() override
+    {
+        new_blocks = std::make_shared<Blocks>();
+        new_blocks_metadata = std::make_shared<BlocksMetadata>();
+        new_hash = std::make_shared<SipHash>();
+    }
+
+    void writeSuffix() override
+    {
+        UInt128 key;
+        String key_str;
+
+        new_hash->get128(key.low, key.high);
+        key_str = key.toHexString();
+
+        Poco::FastMutex::ScopedLock lock(storage.mutex);
+
+        if (storage.getBlocksHashKey() != key_str)
+        {
+            new_blocks_metadata->hash = key_str;
+            new_blocks_metadata->version = storage.getBlocksVersion() + 1;
+
+            for (auto & block : *new_blocks)
+            {
+                block.insert({DataTypeUInt64().createColumnConst(
+                    block.rows(), new_blocks_metadata->version)->convertToFullColumnIfConst(),
+                    std::make_shared<DataTypeUInt64>(),
+                    "_version"});
+            }
+
+            (*storage.blocks_ptr) = new_blocks;
+            (*storage.blocks_metadata_ptr) = new_blocks_metadata;
+
+            storage.condition.broadcast();
+        }
+
+        new_blocks.reset();
+        new_blocks_metadata.reset();
+        new_hash.reset();
+    }
+
+    void write(const Block & block) override
+    {
+        new_blocks->push_back(block);
+        block.updateHash(*new_hash);
+    }
+
+    Block getHeader() const override { return storage.getHeader(); }
+
+private:
+    using SipHashPtr = std::shared_ptr<SipHash>;
+
+    BlocksPtr new_blocks;
+    BlocksMetadataPtr new_blocks_metadata;
+    SipHashPtr new_hash;
+    StorageLiveView & storage;
+};
+
+
+void StorageLiveView::writeIntoLiveView(
+    StorageLiveView & live_view,
+    const Block & block,
+    const Context & context)
+{
+    BlockOutputStreamPtr output = std::make_shared<LiveViewBlockOutputStream>(live_view);
+
+    /// Check if live view has any readers if not
+    /// just reset blocks to empty and do nothing else
+    /// When first reader comes the blocks will be read.
+    {
+        Poco::FastMutex::ScopedLock lock(live_view.mutex);
+        if (!live_view.hasActiveUsers())
+        {
+            live_view.reset();
+            return;
+        }
+    }
+
+    bool is_block_processed = false;
+    BlockInputStreams from;
+    BlocksPtrs mergeable_blocks;
+    BlocksPtr new_mergeable_blocks = std::make_shared<Blocks>();
+
+    {
+        Poco::FastMutex::ScopedLock lock(live_view.mutex);
+
+        mergeable_blocks = live_view.getMergeableBlocks();
+        if (!mergeable_blocks || mergeable_blocks->size() >= context.getGlobalContext().getSettingsRef().max_live_view_insert_blocks_before_refresh)
+        {
+            mergeable_blocks = std::make_shared<std::vector<BlocksPtr>>();
+            BlocksPtr base_mergeable_blocks = std::make_shared<Blocks>();
+            InterpreterSelectQuery interpreter(live_view.getInnerQuery(), context, SelectQueryOptions(QueryProcessingStage::WithMergeableState), Names());
+            auto view_mergeable_stream = std::make_shared<MaterializingBlockInputStream>(
+                interpreter.execute().in);
+            while (Block this_block = view_mergeable_stream->read())
+                base_mergeable_blocks->push_back(this_block);
+            mergeable_blocks->push_back(base_mergeable_blocks);
+            live_view.setMergeableBlocks(mergeable_blocks);
+
+            /// Create from streams
+            for (auto & blocks_ : *mergeable_blocks)
+            {
+                if (blocks_->empty())
+                    continue;
+                auto sample_block = blocks_->front().cloneEmpty();
+                BlockInputStreamPtr stream = std::make_shared<BlocksBlockInputStream>(std::make_shared<BlocksPtr>(blocks_), sample_block);
+                from.push_back(std::move(stream));
+            }
+
+            is_block_processed = true;
+        }
+    }
+
+    if (!is_block_processed)
+    {
+        auto parent_storage = context.getTable(live_view.getSelectDatabaseName(), live_view.getSelectTableName());
+        BlockInputStreams streams = {std::make_shared<OneBlockInputStream>(block)};
+        auto proxy_storage = std::make_shared<ProxyStorage>(parent_storage, std::move(streams), QueryProcessingStage::FetchColumns);
+        InterpreterSelectQuery select_block(live_view.getInnerQuery(),
+            context, proxy_storage,
+            QueryProcessingStage::WithMergeableState);
+        auto data_mergeable_stream = std::make_shared<MaterializingBlockInputStream>(
+            select_block.execute().in);
+        while (Block this_block = data_mergeable_stream->read())
+            new_mergeable_blocks->push_back(this_block);
+
+        if (new_mergeable_blocks->empty())
+            return;
+
+        {
+            Poco::FastMutex::ScopedLock lock(live_view.mutex);
+
+            mergeable_blocks = live_view.getMergeableBlocks();
+            mergeable_blocks->push_back(new_mergeable_blocks);
+
+            /// Create from streams
+            for (auto & blocks_ : *mergeable_blocks)
+            {
+                if (blocks_->empty())
+                    continue;
+                auto sample_block = blocks_->front().cloneEmpty();
+                BlockInputStreamPtr stream = std::make_shared<BlocksBlockInputStream>(std::make_shared<BlocksPtr>(blocks_), sample_block);
+                from.push_back(std::move(stream));
+            }
+        }
+    }
+
+    auto parent_storage = context.getTable(live_view.getSelectDatabaseName(), live_view.getSelectTableName());
+    auto proxy_storage = std::make_shared<ProxyStorage>(parent_storage, std::move(from), QueryProcessingStage::WithMergeableState);
+    InterpreterSelectQuery select(live_view.getInnerQuery(), context, proxy_storage, QueryProcessingStage::Complete);
+    BlockInputStreamPtr data = std::make_shared<MaterializingBlockInputStream>(select.execute().in);
+
+    /// Squashing is needed here because the view query can generate a lot of blocks
+    /// even when only one block is inserted into the parent table (e.g. if the query is a GROUP BY
+    /// and two-level aggregation is triggered).
+    data = std::make_shared<SquashingBlockInputStream>(
+        data, context.getGlobalContext().getSettingsRef().min_insert_block_size_rows, context.getGlobalContext().getSettingsRef().min_insert_block_size_bytes);
+
+    copyData(*data, *output);
+}
+
 
 StorageLiveView::StorageLiveView(
     const String & table_name_,
@@ -259,11 +435,10 @@ void StorageLiveView::noUsersThread(const UInt64 & timeout)
     {
         while (1)
         {
-            Poco::FastMutex::ScopedLock lock(noUsersThreadMutex);
-            if (!noUsersThreadWakeUp && !noUsersThreadCondition.tryWait(noUsersThreadMutex,
-                timeout * 1000))
+            Poco::FastMutex::ScopedLock lock(no_users_thread_mutex);
+            if (!no_users_thread_wakeup && !no_users_thread_condition.tryWait(no_users_thread_mutex, timeout * 1000))
             {
-                noUsersThreadWakeUp = false;
+                no_users_thread_wakeup = false;
                 if (shutdown_called)
                     return;
                 if (hasUsers())
@@ -301,7 +476,7 @@ void StorageLiveView::noUsersThread(const UInt64 & timeout)
 void StorageLiveView::startNoUsersThread(const UInt64 & timeout)
 {
     bool expected = false;
-    if (!startnousersthread_called.compare_exchange_strong(expected, true))
+    if (!start_no_users_thread_called.compare_exchange_strong(expected, true))
         return;
 
     if (is_dropped)
@@ -312,20 +487,20 @@ void StorageLiveView::startNoUsersThread(const UInt64 & timeout)
         if (no_users_thread.joinable())
         {
             {
-                Poco::FastMutex::ScopedLock lock(noUsersThreadMutex);
-                noUsersThreadWakeUp = true;
-                noUsersThreadCondition.signal();
+                Poco::FastMutex::ScopedLock lock(no_users_thread_mutex);
+                no_users_thread_wakeup = true;
+                no_users_thread_condition.signal();
             }
             no_users_thread.join();
         }
         {
-            Poco::FastMutex::ScopedLock lock(noUsersThreadMutex);
-            noUsersThreadWakeUp = false;
+            Poco::FastMutex::ScopedLock lock(no_users_thread_mutex);
+            no_users_thread_wakeup = false;
         }
         if (!is_dropped)
             no_users_thread = std::thread(&StorageLiveView::noUsersThread, this, timeout);
     }
-    startnousersthread_called = false;
+    start_no_users_thread_called = false;
 }
 
 void StorageLiveView::startup()
@@ -341,13 +516,13 @@ void StorageLiveView::shutdown()
 
     if (no_users_thread.joinable())
     {
-        Poco::FastMutex::ScopedLock lock(noUsersThreadMutex);
-        noUsersThreadWakeUp = true;
-        noUsersThreadCondition.signal();
+        Poco::FastMutex::ScopedLock lock(no_users_thread_mutex);
+        no_users_thread_wakeup = true;
+        no_users_thread_condition.signal();
         /// Must detach the no users thread
         /// as we can't join it as it will result
         /// in a deadlock
-        no_users_thread.detach();
+        no_users_thread.detach();   /// TODO Not viable at all.
     }
 }
 
@@ -423,9 +598,9 @@ BlockInputStreams StorageLiveView::watch(
 
         if (no_users_thread.joinable())
         {
-            Poco::FastMutex::ScopedLock lock(noUsersThreadMutex);
-            noUsersThreadWakeUp = true;
-            noUsersThreadCondition.signal();
+            Poco::FastMutex::ScopedLock lock(no_users_thread_mutex);
+            no_users_thread_wakeup = true;
+            no_users_thread_condition.signal();
         }
 
         {
@@ -448,9 +623,9 @@ BlockInputStreams StorageLiveView::watch(
 
         if (no_users_thread.joinable())
         {
-            Poco::FastMutex::ScopedLock lock(noUsersThreadMutex);
-            noUsersThreadWakeUp = true;
-            noUsersThreadCondition.signal();
+            Poco::FastMutex::ScopedLock lock(no_users_thread_mutex);
+            no_users_thread_wakeup = true;
+            no_users_thread_condition.signal();
         }
 
         {
